+++ CHANGES.md
@@ -5,7 +5,9 @@ Features ✨:
  - Handle long click on room in the room list (#395)
 
 Improvements 🙌:
+ - Search reaction by name or keyword in emoji picker
  - Handle code tags (#567)
+ - Support spoiler messages
 
 Other changes:
  - Markdown set to off by default (#412)